add ability to send requests from nyatta

+++ nyatta/src/crypto.h
@@ -8,6 +8,6 @@
 #include "utils.h"
 #include "logger.h"
 
-size_t encrypt(const unsigned char *data, size_t len, const char *private_key_path, const char *public_key_path, unsigned char *ciphertext);
+int encrypt(const unsigned char *data, size_t len, const char *private_key_path, const char *public_key_path, unsigned char *ciphertext);
 
 #endif
